+++ lib/fog/openstack/network.rb
@@ -5,11 +5,11 @@ module Fog
     class OpenStack < Fog::Service
       SUPPORTED_VERSIONS = /v2(\.0)*/
 
-      requires :openstack_auth_url
-      recognizes :openstack_auth_token, :openstack_management_url, :persistent,
-                 :openstack_service_type, :openstack_service_name, :openstack_tenant,
-                 :openstack_api_key, :openstack_username, :openstack_endpoint_type,
-                 :current_user, :current_tenant, :openstack_region
+      requires :auth_url
+      recognizes :auth_token, :management_url, :persistent,
+                 :service_type, :service_name, :tenant,
+                 :api_key, :username, :endpoint_type,
+                 :current_user, :current_tenant, :region
 
       ## MODELS
       #
@@ -202,23 +202,23 @@ module Fog
         end
 
         def initialize(options={})
-          @openstack_username = options[:openstack_username]
-          @openstack_tenant   = options[:openstack_tenant]
+          @username = options[:username]
+          @tenant   = options[:tenant]
         end
 
         def data
-          self.class.data["#{@openstack_username}-#{@openstack_tenant}"]
+          self.class.data["#{@username}-#{@tenant}"]
         end
 
         def reset_data
-          self.class.data.delete("#{@openstack_username}-#{@openstack_tenant}")
+          self.class.data.delete("#{@username}-#{@tenant}")
         end
 
         def credentials
           { :provider                 => 'openstack',
-            :openstack_auth_url       => @openstack_auth_uri.to_s,
-            :openstack_auth_token     => @auth_token,
-            :openstack_management_url => @openstack_management_url }
+            :auth_url       => @auth_uri.to_s,
+            :auth_token     => @auth_token,
+            :management_url => @management_url }
         end
       end
 
@@ -227,26 +227,26 @@ module Fog
         attr_reader :current_tenant
 
         def initialize(options={})
-          @openstack_auth_token = options[:openstack_auth_token]
+          @auth_token = options[:auth_token]
 
-          unless @openstack_auth_token
+          unless @auth_token
             missing_credentials = Array.new
-            @openstack_api_key  = options[:openstack_api_key]
-            @openstack_username = options[:openstack_username]
+            @api_key  = options[:api_key]
+            @username = options[:username]
 
-            missing_credentials << :openstack_api_key  unless @openstack_api_key
-            missing_credentials << :openstack_username unless @openstack_username
+            missing_credentials << :api_key  unless @api_key
+            missing_credentials << :username unless @username
             raise ArgumentError, "Missing required arguments: #{missing_credentials.join(', ')}" unless missing_credentials.empty?
           end
 
-          @openstack_tenant               = options[:openstack_tenant]
-          @openstack_auth_uri             = URI.parse(options[:openstack_auth_url])
-          @openstack_management_url       = options[:openstack_management_url]
-          @openstack_must_reauthenticate  = false
-          @openstack_service_type         = options[:openstack_service_type] || ['network']
-          @openstack_service_name         = options[:openstack_service_name]
-          @openstack_endpoint_type        = options[:openstack_endpoint_type] || 'publicURL'
-          @openstack_region               = options[:openstack_region]
+          @tenant               = options[:tenant]
+          @auth_uri             = URI.parse(options[:auth_url])
+          @management_url       = options[:management_url]
+          @must_reauthenticate  = false
+          @service_type         = options[:service_type] || ['network']
+          @service_name         = options[:service_name]
+          @endpoint_type        = options[:endpoint_type] || 'publicURL'
+          @region               = options[:region]
 
           @connection_options = options[:connection_options] || {}
 
@@ -261,12 +261,12 @@ module Fog
 
         def credentials
           { :provider                 => 'openstack',
-            :openstack_auth_url       => @openstack_auth_uri.to_s,
-            :openstack_auth_token     => @auth_token,
-            :openstack_management_url => @openstack_management_url,
+            :auth_url       => @auth_uri.to_s,
+            :auth_token     => @auth_token,
+            :management_url => @management_url,
             :current_user             => @current_user,
             :current_tenant           => @current_tenant,
-            :openstack_region         => @openstack_region }
+            :region         => @region }
         end
 
         def reload
@@ -285,7 +285,7 @@ module Fog
             }))
           rescue Excon::Errors::Unauthorized => error
             if error.response.body != 'Bad username or password' # token expiration
-              @openstack_must_reauthenticate = true
+              @must_reauthenticate = true
               authenticate
               retry
             else # bad credentials
@@ -308,17 +308,17 @@ module Fog
         private
 
         def authenticate
-          if !@openstack_management_url || @openstack_must_reauthenticate
+          if !@management_url || @must_reauthenticate
             options = {
-              :openstack_tenant   => @openstack_tenant,
-              :openstack_api_key  => @openstack_api_key,
-              :openstack_username => @openstack_username,
-              :openstack_auth_uri => @openstack_auth_uri,
-              :openstack_auth_token => @openstack_auth_token,
-              :openstack_service_type => @openstack_service_type,
-              :openstack_service_name => @openstack_service_name,
-              :openstack_endpoint_type => @openstack_endpoint_type,
-              :openstack_region => @openstack_region
+              :tenant   => @tenant,
+              :api_key  => @api_key,
+              :username => @username,
+              :auth_uri => @auth_uri,
+              :auth_token => @auth_token,
+              :service_type => @service_type,
+              :service_name => @service_name,
+              :endpoint_type => @endpoint_type,
+              :region => @region
             }
 
             credentials = Fog::OpenStack.authenticate_v2(options, @connection_options)
@@ -326,13 +326,13 @@ module Fog
             @current_user = credentials[:user]
             @current_tenant = credentials[:tenant]
 
-            @openstack_must_reauthenticate = false
+            @must_reauthenticate = false
             @auth_token = credentials[:token]
-            @openstack_management_url = credentials[:server_management_url]
-            uri = URI.parse(@openstack_management_url)
+            @management_url = credentials[:server_management_url]
+            uri = URI.parse(@management_url)
           else
-            @auth_token = @openstack_auth_token
-            uri = URI.parse(@openstack_management_url)
+            @auth_token = @auth_token
+            uri = URI.parse(@management_url)
           end
 
           @host   = uri.host